+++ gen/aws/generators/paramsdoc.go
@@ -143,6 +143,7 @@ func loadAllRefs() map[string]string {
 		filepath.Join("ecr", "2015-09-21", "docs-2.json"),
 		filepath.Join("ecs", "2014-11-13", "docs-2.json"),
 		filepath.Join("application-autoscaling", "2016-02-06", "docs-2.json"),
+		filepath.Join("acm", "2015-12-08", "docs-2.json"),
 	}
 
 	entriesC := make(chan *entries)