+++ template/revert.go
@@ -105,6 +105,8 @@ func (te *Template) Revert() (*Template, error) {
 				case "database":
 					params = append(params, fmt.Sprintf("id=%s", quoteParamIfNeeded(cmd.CmdResult)))
 					params = append(params, "skip-snapshot=true")
+				case "certificate":
+					params = append(params, fmt.Sprintf("arn=%s", quoteParamIfNeeded(cmd.CmdResult)))
 				case "policy":
 					params = append(params, fmt.Sprintf("arn=%s", quoteParamIfNeeded(cmd.CmdResult)))
 				case "queue":